requirements.txt
requests==2.25.1

google-api-python-client==2.123.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

azure-identity==1.15.0
msal==1.28.0
msal-extensions==1.1.0
